+++ packages/gasket-plugin-docusaurus/lib/docs-view.js
@@ -2,6 +2,7 @@ const defaultsDeep = require('lodash.defaultsdeep');
 const { existsSync } = require('fs');
 const { writeFile } = require('fs').promises;
 const path = require('path');
+const { tryRequire } = require('@gasket/utils');
 const generateDefaultConfig = require('./generate-default-config');
 const pluginConfigFile = 'docusaurus.config.js';
 const defaultConfig = {
@@ -9,7 +10,16 @@ const defaultConfig = {
   host: 'localhost'
 };
 
+function checkDevDependencies() {
+  const preset = tryRequire('@docusaurus/preset-classic');
+  const core = tryRequire('@docusaurus/core');
+  if (!preset || !core) {
+    throw new Error('Missing devDependencies. Please run `npm i -D @docusaurus/core @docusaurus/preset-classic`');
+  }
+}
+
 module.exports = async function docsView(gasket) {
+  checkDevDependencies();
   const { start } = require('@docusaurus/core/lib');
   const { config } = gasket;
   const { name } = gasket.metadata.app;