+++ packages/gasket-plugin-docusaurus/CHANGELOG.md
@@ -1,5 +1,13 @@
 # `@gasket/plugin-docusaurus`
 
+### 6.44.5
+
+- add docusaurus to devDeps, add dep check, add create hook ([#658])
+
+### 6.44.3
+
+- Lazy load Docusaurus, update devDependencies ([#650])
+
 ### 6.35.0
 
 - Upgrade @docusaurus/* to v2.2.0 ([#481])
@@ -34,3 +42,5 @@
 [#442]: https://github.com/godaddy/gasket/pull/442
 [#460]: https://github.com/godaddy/gasket/pull/460
 [#481]: https://github.com/godaddy/gasket/pull/481
+[#650]: https://github.com/godaddy/gasket/pull/650
+[#658]: https://github.com/godaddy/gasket/pull/658