+++ packages/gasket-plugin-express/CHANGELOG.md
@@ -1,6 +1,10 @@
 # `@gasket/plugin-express`
 
-- Add a logger to the request object
+### 6.44.0
+
+- Documentation & type fixes ([#637])
+  - Add missing type definition for `routes` config property
+  - Add `middlewareInclusionRegex` config property, deprecating `excludedRoutesRegex`
 
 ### 6.41.2
 
@@ -63,7 +67,6 @@
 - Separate express plugins from core
 - Initial release.
 
-
 [#23]: https://github.com/godaddy/gasket/pull/23
 [#181]: https://github.com/godaddy/gasket/pull/181
 [#247]: https://github.com/godaddy/gasket/pull/247
@@ -76,3 +79,4 @@
 [#442]: https://github.com/godaddy/gasket/pull/442
 [#460]: https://github.com/godaddy/gasket/pull/460
 [#613]: https://github.com/godaddy/gasket/pull/613
+[#637]: https://github.com/godaddy/gasket/pull/637